+++ src/components/common/Carousel/carousel.tsx
@@ -1,8 +1,8 @@
-import { ReactJSXElement } from '@emotion/react/types/jsx-namespace';
+import type { ReactJSXElement } from '@emotion/react/types/jsx-namespace';
 import { AnimatePresence, motion } from 'framer-motion';
 import React, { useState } from 'react';
 
-import { TabNavMenu } from '../../navigation/tabNavMenu/tabNavMenu';
+import { TabNavMenu } from '@/components/navigation/tabNavMenu/tabNavMenu';
 
 interface CarouselProps {
   names: string[] | string;
