+++ src/components/navigation/topMenu/topMenu.tsx
@@ -5,8 +5,8 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import React, { useState } from 'react';
 
-import Background from '../../../../public/background.png';
-import SearchBar from '../../common/SearchBar/searchBar';
+import Background from '@/../public/background.png';
+import SearchBar from '@/components/common/SearchBar/searchBar';
 
 /**
  * This is a component to hold UTD Trends branding and basic navigation
