+++ src/pages/_app.tsx
@@ -13,9 +13,9 @@ import Head from 'next/head';
 import { useRouter } from 'next/router';
 import React from 'react';
 
-import tailwindConfig from '../../tailwind.config.js';
-import FeedbackPopup from '../components/common/FeedbackPopup/feedbackPopup';
-import GitHubButton from '../components/common/GitHubButton/gitHubButton';
+import tailwindConfig from '@/../tailwind.config.js';
+import FeedbackPopup from '@/components/common/FeedbackPopup/feedbackPopup';
+import GitHubButton from '@/components/common/GitHubButton/gitHubButton';
 
 const inter = Inter({
   subsets: ['latin'],
