+++ src/pages/_document.tsx
@@ -1,7 +1,7 @@
 import { Head, Html, Main, NextScript } from 'next/document';
 import React from 'react';
 
-import tailwindConfig from '../../tailwind.config.js';
+import tailwindConfig from '@/../tailwind.config.js';
 
 function Document() {
   return (
