+++ src/scripts/generateCombosTable.ts
@@ -3,9 +3,11 @@ Build the combos table
 */
 import * as fs from 'fs';
 
-import aggregatedData from '../data/autocomplete_data.json';
-import SearchQuery from '../modules/SearchQuery/SearchQuery';
-import searchQueryEqual from '../modules/searchQueryEqual/searchQueryEqual';
+import aggregatedData from '@/data/autocomplete_data.json';
+import {
+  type SearchQuery,
+  searchQueryEqual,
+} from '@/modules/SearchQuery/SearchQuery';
 
 export type TableType = { [key: string]: SearchQuery[] };
