+++ tsconfig.json
@@ -18,8 +18,11 @@
       // remap rsg-components/anything to its version in react-styleguidist
       "rsg-components/*": [
         "./node_modules/react-styleguidist/lib/client/rsg-components/*"
-      ]
-    }
+      ],
+      "@/*": ["*"]
+    },
+    "baseUrl": "src",
+    "verbatimModuleSyntax": true
   },
   "include": ["src", "next-env.d.ts", "**/*.ts", "**/*.tsx", "/**/**/*/.tsx"],
   "exclude": ["node_modules"],
